article-to-TTS-converter/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "article2tts"
version = "0.1.0"
description = "Converts academic articles (PDF/Word) to TTS-friendly Markdown for Obsidian"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "langdetect>=1.0.9",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
article2tts = "article2tts.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
